requirements.txt
fastapi
uvicorn
sqlalchemy
openai
python-dotenv
fpdf2
jinja2
python-multipart
httpx
beautifulsoup4


